fix: SP_PLAYER uint8→uint16 for Spongent-256 (indices up to 271)
Spongent-256 width=272 produces pLayer indices up to 271, which
overflows uint8. Changed SP_PLAYER to uint16_t in the CUDA kernel
and struct.pack('<NH', ...) in the host-side _compute_sp_player().

+++ src/ggm/kernels/spongent_kernel.cu
@@ -30,7 +30,7 @@
 extern "C" {
 
 __constant__ uint8_t SP_SBOX[16];
-__constant__ uint8_t SP_PLAYER[SP_WIDTH];
+__constant__ uint16_t SP_PLAYER[SP_WIDTH];
 
 }
